Fix circular references
Removed /ui file references to /events scripts
instead all bindings occur in app.js
