+++ assets/scripts/auth/events.js
@@ -4,46 +4,35 @@
 // Controller for Authentication requests
 
 const authApi = require('./api')
-const ui = require('./ui')
+const authUi = require('./ui')
 const getFormFields = require('../../../lib/get-form-fields')
-const store = require('../store')
-const utils = require('../main/utility')
 
 const onLoginSubmit = (event) => {
   const formData = getFormData(event)
-  // authApi.signIn(formData)
-  //   .then(authUi.loginSuccess)
-  //   .catch(authUi.authFail)
-  // if authUi.loginSuccess not working,
   authApi.signIn(formData)
-    .then((responseData) => {
-      const msg = 'Login successful'
-      console.log(msg)
-      store.user = responseData.user
-      utils.userMessage(msg)
-      $('#userEmail').text(store.user.email)
-    }).catch(ui.authFail)
+    .then(authUi.loginSuccess)
+    .catch(authUi.authFail)
 }
 
 const onSignupSubmit = (event) => {
   const formData = getFormData(event)
   authApi.signUp(formData)
-    .then(ui.signUpSuccess)
-    .catch(ui.authFail)
+    .then(authUi.signUpSuccess)
+    .catch(authUi.authFail)
 }
 
 const onChangePasswordSubmit = (event) => {
   const formData = getFormData(event)
   authApi.changePassword(formData)
-    .then(ui.changePasswordSuccess)
-    .catch(ui.authenticationError)
+    .then(authUi.changePasswordSuccess)
+    .catch(authUi.authenticationError)
 }
 
 const onSignoutConfirm = (event) => {
   event.preventDefault()
   authApi.signOut()
-    .then(ui.signOutSuccess)
-    .catch(ui.authenticationError)
+    .then(authUi.signOutSuccess)
+    .catch(authUi.authenticationError)
 }
 
 const getFormData = (event) => {
