+++ assets/scripts/auth/ui.js
@@ -12,7 +12,7 @@
  *****************************/
 
 const store = require('../store')
-const authEvents = require('./events')
+// const authEvents = require('./events') // circular reference
 const utils = require('../main/utility')
 
 // Public: check for token, update all elements
@@ -29,52 +29,7 @@ const refreshAuthElements = () => {
 
 const clearUserData = () => {
   $('#userEmail').html('')
-  $('#playerStats').html('')
-}
-
-// Public: set fields and target of modal auth form; show form
-const modalForm = (event) => {
-  const btn = $(event.target)
-  const currentOp = btn.data('operation')
-  $('#authOperation').val(currentOp)
-  $('#authenticationForm .form-group').hide()
-
-  let target, title
-  switch (currentOp) {
-    case 'login':
-      target = authEvents.onLoginSubmit
-      $('#emailFieldEnclosure').show()
-      $('#passwordFieldEnclosure').show()
-      title = 'Login'
-      break
-    case 'signup':
-      target = authEvents.onSignupSubmit
-      $('#emailFieldEnclosure').show()
-      $('#passwordFieldEnclosure').show()
-      $('#passwordConfFieldEnclosure').show()
-      title = 'Sign Up'
-      break
-    case 'changepw':
-      target = authEvents.onChangePasswordSubmit
-      $('#oldPasswordFieldEnclosure').show()
-      $('#newPasswordFieldEnclosure').show()
-      title = 'Change Password'
-      break
-    case 'signout':
-      target = authEvents.onSignoutConfirm
-      $('#signoutConfirmEnclosure').show()
-      title = 'Sign Out'
-  }
-  $('#authModalTitle').text(title)
-  $('#authenticationForm').on('submit', target)
-
-  // make this button submit the modal form and also close the modal
-  $('#authSubmitBtn').on('click', () => {
-    $('#authenticationForm').submit()
-    $('#authModal').modal('hide')
-  })
-
-  $('#authModal').modal('show')
+  // Clear stats
 }
 
 // Public
@@ -86,7 +41,6 @@ const authFail = () => {
 // Public
 const loginSuccess = (responseData) => {
   const msg = 'Authentication successful'
-  console.log(msg)
   store.user = responseData.user
   utils.userMessage(msg)
   $('#userEmail').text(store.user.email)
@@ -118,7 +72,6 @@ const signOutSuccess = () => {
 
 module.exports = {
   refreshAuthElements,
-  modalForm,
   authFail,
   authenticationError,
   loginSuccess,
